backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
httpx==0.28.1
python-jose[cryptography]==3.3.0
cryptography==44.0.0
pydantic-settings==2.6.1
anthropic==0.40.0
apscheduler==3.10.4
python-multipart==0.0.17
pytz==2024.2
